improved bad view tests and added more postgre test cases.  Postgress ListTables now works for views.

+++ Tests/FAnsiTests/Database/DiscoverTablesTests.cs
@@ -13,6 +13,7 @@ namespace FAnsiTests.Database
         [TestCase(DatabaseType.MySql)] 
         [TestCase(DatabaseType.Oracle)]
         [TestCase(DatabaseType.MicrosoftSQLServer)]
+        [TestCase(DatabaseType.PostgreSql)]
         public void Test_DiscoverTables_Normal(DatabaseType dbType)
         {
             var db = GetTestDatabase(dbType);
@@ -45,6 +46,7 @@ namespace FAnsiTests.Database
         [TestCase(DatabaseType.MySql)] 
         [TestCase(DatabaseType.Oracle)]
         [TestCase(DatabaseType.MicrosoftSQLServer)]
+        [TestCase(DatabaseType.PostgreSql)]
         public void Test_DiscoverTables_WithInvalidNames_Skipped(DatabaseType dbType)
         {
             var db = GetTestDatabase(dbType);
@@ -73,7 +75,7 @@ namespace FAnsiTests.Database
             Assert.AreEqual(1,tbls.Length);
             Assert.AreEqual(1, tbls.Count(t => t.GetRuntimeName().Equals("FF",StringComparison.CurrentCultureIgnoreCase)));
             
-            DropBadTable(db);
+            DropBadTable(db,false);
         }
 
         /// <summary>
@@ -83,6 +85,7 @@ namespace FAnsiTests.Database
         [TestCase(DatabaseType.MySql)] 
         [TestCase(DatabaseType.Oracle)]
         [TestCase(DatabaseType.MicrosoftSQLServer)]
+        [TestCase(DatabaseType.PostgreSql)]
         public void Test_DiscoverViews_WithInvalidNames_Skipped(DatabaseType dbType)
         {
             var db = GetTestDatabase(dbType);
@@ -115,10 +118,10 @@ namespace FAnsiTests.Database
             Assert.AreEqual(0,tbls.Count(t=>t.TableType == TableType.View));
             Assert.AreEqual(1, tbls.Count(t => t.GetRuntimeName().Equals("FF",StringComparison.CurrentCultureIgnoreCase)));
             
-            DropBadView(db);
+            DropBadView(db,false);
         }
 
-        private void DropBadTable(DiscoveredDatabase db)
+        private void DropBadTable(DiscoveredDatabase db,bool ignoreFailure)
         {
             using(var con = db.Server.GetConnection())
             {
@@ -130,6 +133,9 @@ namespace FAnsiTests.Database
                 }
                 catch (Exception)
                 {
+                    if (!ignoreFailure)
+                        throw;
+
                     Console.WriteLine("Drop table failed, this is expected, since FAnsi won't see this dodgy table name we can't drop it as normal before tests");
                 }
             }
@@ -145,6 +151,8 @@ namespace FAnsiTests.Database
                     return "`BB (ff)`";
                 case DatabaseType.Oracle:
                     return  db.GetRuntimeName() + ".\"BB (ff)\"";
+                case DatabaseType.PostgreSql:
+                    return  '"' + db.GetRuntimeName() + "\".public.\"BB (ff)\"";
                 default:
                     throw new ArgumentOutOfRangeException(nameof(db.Server.DatabaseType), db.Server.DatabaseType, null);
             }
@@ -153,7 +161,7 @@ namespace FAnsiTests.Database
         private void CreateBadTable(DiscoveredDatabase db)
         {
             //drop it if it exists
-            DropBadTable(db);
+            DropBadTable(db,true);
 
             using(var con = db.Server.GetConnection())
             {
@@ -164,12 +172,8 @@ namespace FAnsiTests.Database
         }
 
 
-        private void DropBadView(DiscoveredDatabase db)
+        private void DropBadView(DiscoveredDatabase db, bool ignoreFailure)
         {
-            //the table that the view reads from
-            var abc = db.ExpectTable("ABC");
-            if(abc.Exists())
-                abc.Drop();
             
             using(var con = db.Server.GetConnection())
             {
@@ -181,23 +185,35 @@ namespace FAnsiTests.Database
                 }
                 catch (Exception)
                 {
+                    if (!ignoreFailure)
+                        throw;
+
                     Console.WriteLine("Drop view failed, this is expected, since FAnsi won't see this dodgy table name we can't drop it as normal before tests");
                 }
             }
+
+            //the table that the view reads from
+            var abc = db.ExpectTable("ABC");
+            if(abc.Exists())
+                abc.Drop();
+
         }
 
 
         private void CreateBadView(DiscoveredDatabase db)
         {
             //drop it if it exists
-            DropBadView(db);
+            DropBadView(db,true);
 
             db.CreateTable("ABC",new DatabaseColumnRequest[]{new DatabaseColumnRequest("A",new DatabaseTypeRequest(typeof(int)))});
 
             using(var con = db.Server.GetConnection())
             {
                 con.Open();
-                var cmd = db.Server.GetCommand($"CREATE VIEW {GetBadTableName(db)} as select * from ABC",con);
+ 
+                var viewname = db.Server.GetQuerySyntaxHelper().EnsureWrapped("ABC");
+
+                var cmd = db.Server.GetCommand($"CREATE VIEW {GetBadTableName(db)} as select * from " + viewname,con);
                 cmd.ExecuteNonQuery();
             }
         }
